+++ orchestration/orchestrator.py
@@ -40,8 +40,8 @@ from torch import nn
 from torch.utils.data import DataLoader
 from torchvision import datasets, transforms
 
-from model_registry import get_model_spec
-from train_env import apply_seed
+from .model_registry import get_model_spec
+from .train_env import apply_seed
 
 console = Console()
