+++ trainers/fastervit.py
@@ -40,7 +40,7 @@ from torch.utils.data import DataLoader
 from torchvision import datasets, transforms
 from torchvision.transforms import InterpolationMode
 
-from train_env import (
+from orchestration.train_env import (
     apply_seed,
     create_console,
     env_int,
